Split the functional groups into species and group

Split strings containing functional groups into species and group names. Functional groups can contain age ranges, plus groups, or sub categories, e.g., juvenile. This information is extracted from the functional group name and returned as the group name.

split_functional_groups(x)

Arguments

x

A character vector of functional group names. This is most often created by running unique() on the functional group column of the data.

Value

A tibble with the following three columns: functional_group, species, and group, is returned.

Author

Kelli F. Johnson

Examples

# A hypothetical example of functional groups
split_functional_groups(
  c("spiny dogfish 0", "spiny dogfish 1-2", "spiny dogfish +")
)
#> # A tibble: 3 × 3
#>   functional_group  species       group
#>   <chr>             <chr>         <chr>
#> 1 spiny dogfish 0   spiny dogfish 0    
#> 2 spiny dogfish 1-2 spiny dogfish 1-2  
#> 3 spiny dogfish +   spiny dogfish +    
split_functional_groups(c("spiny dogfish", "spiny dogfish juvenile"))
#> # A tibble: 2 × 3
#>   functional_group       species       group   
#>   <chr>                  <chr>         <chr>   
#> 1 spiny dogfish          spiny         dogfish 
#> 2 spiny dogfish juvenile spiny dogfish juvenile
data("ewe_nwatlantic_base", package = "ecosystemdata")
split_functional_groups(unique(ewe_nwatlantic_base[["functional_group"]]))
#> # A tibble: 22 × 3
#>    functional_group species      group
#>    <chr>            <chr>        <chr>
#>  1 striped bass 0   striped bass 0    
#>  2 striped bass 2-5 striped bass 2-5  
#>  3 striped bass 6+  striped bass 6+   
#>  4 menhaden 0       menhaden     0    
#>  5 menhaden 1       menhaden     1    
#>  6 menhaden 2       menhaden     2    
#>  7 menhaden 3       menhaden     3    
#>  8 menhaden 4       menhaden     4    
#>  9 menhaden 5       menhaden     5    
#> 10 menhaden 6+      menhaden     6+   
#> # ℹ 12 more rows